[feature] add multiple context feature #6
## WHY:
Currently, when changing Site to test (eg. from dev to stg env), you have to fill in the form again and again. As a result, it takes a lot of time just to switch the context.

## WHAT:
The PR makes it possible to keep the current form state for later usage when switching the context for testing something else. As a first step, there are 5 possible contexts to use.

## HOW:
It prefixes keys stored in LocalStorage with the context name (numbers from 1 to 5)

![2023-03-13 13 16 49](https://user-images.githubusercontent.com/14330744/224699819-eea3f23a-a3f6-488b-8439-2fde2c49fa23.gif)